// web/src/lib/api/model/artifactKind.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Silo
 * OpenAPI spec version: 0.1.0
 */

export type ArtifactKind = (typeof ArtifactKind)[keyof typeof ArtifactKind]

export const ArtifactKind = {
  installer: "installer",
  patch: "patch",
  language_pack: "language_pack",
  disc: "disc",
  rom: "rom",
  extra: "extra",
  save: "save",
  manual: "manual",
  other: "other",
} as const